LayoutTests/media/loopend-limits.html


<video loopend=1.5></video>
<audio loopend=1.5></audio>

<video end=1.0></video>
<audio end=1.0></audio>

<video ></video>
<audio ></audio>

<video loopend="s1.0"></video>
<audio loopend="s1.0"></audio>

<video loopend='s1.0' end=1.0></video>
<audio loopend='s1.0' end=1.0></audio>


<p>Test loopend attribute parsing, limits, and defaults.</p>

<script src=video-test.js></script>

<script>

// 4.8.10.5. Offsets into the media resource
// loopend: default is the value of the end DOM attribute.
// end: default value is infinity

consoleWrite("<em>&lt;video loopend=1.5&gt;</em> - loopend specified, should not change");
var video = document.getElementsByTagName('video')[0];
var audio = document.getElementsByTagName('audio')[0];
testExpected("video.loopEnd", 1.5);
testExpected("audio.loopEnd", 1.5);

consoleWrite("<br><em>&lt;video end=1.0&gt;</em> - loopend not specified, should return end attr");
var video = document.getElementsByTagName('video')[1];
var audio = document.getElementsByTagName('audio')[1];
testExpected("video.loopEnd", 1.0);
testExpected("audio.loopEnd", 1.0);

consoleWrite("<br><em>&lt;video &gt;</em> - no loopend and no end, default should be infinity (end default)");
var video = document.getElementsByTagName('video')[2];
var audio = document.getElementsByTagName('audio')[2];
testExpected("video.loopEnd", Number.POSITIVE_INFINITY);
testExpected("audio.loopEnd", Number.POSITIVE_INFINITY);

consoleWrite("<br><em>&lt;video loopend='s1.0'&gt;</em> - bogus loopend and no end, default should be infinity (end default)");
var video = document.getElementsByTagName('video')[3];
var audio = document.getElementsByTagName('audio')[3];
testExpected("video.loopEnd", Number.POSITIVE_INFINITY);
testExpected("audio.loopEnd", Number.POSITIVE_INFINITY);

consoleWrite("<br><em>&lt;video loopend='s1.0' end=1.0&gt;</em> - bogus loopend, default should be end attr");
var video = document.getElementsByTagName('video')[4];
var audio = document.getElementsByTagName('audio')[4];
testExpected("video.loopEnd", 1.0);
testExpected("audio.loopEnd", 1.0);

endTest()
</script>
